Update fswatcher.config
Updated the FILE_LOGGING parameter default from 'false' to 'true'. By default, the parameter is commented out thus disabling the file logging. Shipping the parameter as 'false' by default is redundant. If the default is set to 'true', the user only needs to uncomment the line and logging will be enabled when initially configuring the application.